requirements.txt
flask>=3.0
requests>=2.31
feedparser>=6.0
newspaper3k>=0.2.8
readability-lxml>=0.8
beautifulsoup4>=4.12
lxml>=5.0
